requirements.txt
MySQL-python==1.2.5
Django==1.8.3
docutils==0.12
pytz==2015.4
django-admin-sortable2==0.5.2
django-mptt==0.7.4
feedparser==5.2.1
django-debug-toolbar==1.3.2
django-jet==0.0.4